HW-1 Movie 데이터 클래스의 userRating멤버의 타입을 Double에서 float로 변경

+++ 172900/GoogleArchitectureStudy/app/src/main/java/com/onit/googlearchitecturestudy/Movie.kt
@@ -3,7 +3,7 @@ package com.onit.googlearchitecturestudy
 data class Movie(
     val posterImage: String,
     val title: String,
-    val userRating: Double,
+    val userRating: Float,
     val releaseDate: String,
     val director: String,
     val actor: String,
